Remove 'Custom' language version. Unify language settings panels.
