+++ ILSpy/DecompilationOptions.cs
@@ -18,7 +18,6 @@
 
 using System;
 using System.Threading;
-using ICSharpCode.Decompiler;
 using ICSharpCode.ILSpy.Options;
 
 namespace ICSharpCode.ILSpy
@@ -68,43 +67,24 @@ namespace ICSharpCode.ILSpy
 		internal bool IsDebug = false;
 
 		public DecompilationOptions()
-			: this(MainWindow.Instance.CurrentLanguageVersion, DecompilerSettingsPanel.CurrentDecompilerSettings, DisplaySettingsPanel.CurrentDisplaySettings, CSharpDecompilerSettingsPanel.CurrentCSharpSettings)
+			: this(MainWindow.Instance.CurrentLanguageVersion, DecompilerSettingsPanel.CurrentDecompilerSettings, DisplaySettingsPanel.CurrentDisplaySettings)
 		{
 		}
 
 		public DecompilationOptions(LanguageVersion version)
-			: this(version, DecompilerSettingsPanel.CurrentDecompilerSettings, DisplaySettingsPanel.CurrentDisplaySettings, CSharpDecompilerSettingsPanel.CurrentCSharpSettings)
+			: this(version, DecompilerSettingsPanel.CurrentDecompilerSettings, DisplaySettingsPanel.CurrentDisplaySettings)
 		{
 		}
 
-		public DecompilationOptions(LanguageVersion version, Options.DecompilerSettings settings, Options.DisplaySettings displaySettings,  Decompiler.DecompilerSettings csharpSettings)
+		public DecompilationOptions(LanguageVersion version, Decompiler.DecompilerSettings settings, Options.DisplaySettings displaySettings)
 		{
-			if (version.Version == "Custom") {
-				var newSettings = this.DecompilerSettings = csharpSettings.Clone();
-				newSettings.AlwaysUseBraces = settings.AlwaysUseBraces;
-				newSettings.ExpandMemberDefinitions = displaySettings.ExpandMemberDefinitions;
-				newSettings.FoldBraces = displaySettings.FoldBraces;
-				newSettings.RemoveDeadCode = settings.RemoveDeadCode;
-				newSettings.ShowDebugInfo = settings.ShowDebugInfo;
-				newSettings.ShowXmlDocumentation = settings.ShowXmlDocumentation;
-				newSettings.UseDebugSymbols = settings.UseDebugSymbols;
-				newSettings.UsingDeclarations = settings.UsingDeclarations;
-				newSettings.ApplyWindowsRuntimeProjections = settings.ApplyWindowsRuntimeProjections;
-			} else {
-				if (!Enum.TryParse(version.Version, out Decompiler.CSharp.LanguageVersion languageVersion)) 
-					languageVersion = Decompiler.CSharp.LanguageVersion.Latest;
-				this.DecompilerSettings = new Decompiler.DecompilerSettings(languageVersion) {
-					AlwaysUseBraces = settings.AlwaysUseBraces,
-					ExpandMemberDefinitions = displaySettings.ExpandMemberDefinitions,
-					FoldBraces = displaySettings.FoldBraces,
-					RemoveDeadCode = settings.RemoveDeadCode,
-					ShowDebugInfo = settings.ShowDebugInfo,
-					ShowXmlDocumentation = settings.ShowXmlDocumentation,
-					UseDebugSymbols = settings.UseDebugSymbols,
-					UsingDeclarations = settings.UsingDeclarations,
-					ApplyWindowsRuntimeProjections = settings.ApplyWindowsRuntimeProjections,
-				};
-			}
+			if (!Enum.TryParse(version.Version, out Decompiler.CSharp.LanguageVersion languageVersion)) 
+				languageVersion = Decompiler.CSharp.LanguageVersion.Latest;
+			var newSettings = this.DecompilerSettings = settings.Clone();
+			newSettings.SetLanguageVersion(languageVersion);
+			newSettings.ExpandMemberDefinitions = displaySettings.ExpandMemberDefinitions;
+			newSettings.FoldBraces = displaySettings.FoldBraces;
+			newSettings.ShowDebugInfo = displaySettings.ShowDebugInfo;
 		}
 	}
 }
